fix: drop the patches that republished private avatars and community images
The upload path already stores avatars and community images as public, so
these patches only ever mattered for files uploaded before that landed. On
production there are none: every GP User Profile.image and GP Team.image is
already public, so both patches were no-ops.

Keeping them was not free. Publishing a file is not a local decision. Saving a
File with is_private = 0 runs handle_is_private_changed, which calls
update_existing_file_docs, and that bulk update sets file_url and is_private on
every File row sharing the published file's content_hash. Propagation keys on
content_hash, not on file_url, so the allowed_owners check added in d512f379
could not hold: it inspects the rows at one URL and the update reaches rows it
never looked at.

Ordinary dedup is enough to create the shared-hash condition. File.save_file
dedups on (content_hash, is_private), so two users who independently upload the
same bytes privately end up with two rows, two owners and one content_hash. No
attacker is involved. On production 291 private File rows sit in 128 such
cross-owner hash groups, and they are attachments on discussions, comments,
pages and drafts, which is exactly the content that is meant to stay private.

Removing the migration removes the whole surface rather than guarding it. The
cost is that a site with old private avatars keeps them private until someone
uploads a new one, which is a broken image in a digest email at worst.

Reported upstream as https://github.com/frappe/frappe/issues/41667.

attach_profile_bento_card_images is unrelated and stays. The set_image
ownership guard also stays, since it sits on the upload path rather than the
migration; its tests move to test_profile_image_ownership.py, which is what the
file now covers.
